requirements.txt
openai-whisper
torch
torchaudio
transformers
sentencepiece
datasets
soundfile
librosa
tqdm
jiwer
TTS
gradio
